+++ packages/components-react/src/index.ts
@@ -83,9 +83,25 @@ export { HeadingGroup, type HeadingGroupProps } from './HeadingGroup';
 export { Hero, type HeroProps } from './Hero';
 export { Icon, type IconProps, type RHCIconID, getIconSet, registerIconSet } from './Icon';
 export { Image, type ImageProps } from './Image';
+export {
+  LanguageNavigation,
+  type LanguageNavigationContentProps,
+  type LanguageNavigationItemProps,
+  type LanguageNavigationRootProps,
+  type LanguageNavigationTriggerProps,
+} from './LanguageNavigation';
 export { Link, type LinkProps } from './Link';
+export { LinkButton, type LinkButtonProps } from './LinkButton';
 export { LinkList, LinkListLink, type LinkListLinkProps, type LinkListProps } from './LinkList';
 export { LinkListCard } from './LinkListCard';
+export {
+  Listbox,
+  ListboxOption,
+  ListboxOptionGroup,
+  type ListboxOptionGroupProps,
+  type ListboxOptionProps,
+  type ListboxProps,
+} from './Listbox';
 export { Logo, type LogoProps } from './Logo';
 export { MessageList, type MessageListProps } from './MessageList';
 export { MessageListItem, type MessageListItemProps } from './MessageListItem';