+++ packages/storybook/src/components-react/listbox.stories.tsx
@@ -1,0 +1,34 @@
+import { Listbox, ListboxOption } from '@rijkshuisstijl-community/components-react';
+import { Meta, StoryObj } from '@storybook/react-vite';
+
+const meta = {
+  title: 'Listbox',
+  id: 'rhc-listbox',
+  component: Listbox,
+  args: {
+    children: 'Label',
+    disabled: false,
+    'aria-label': 'Listbox example',
+  },
+  parameters: {
+    github:
+      'https://github.com/nl-design-system/rijkshuisstijl-community/blob/main/packages/components-react/src/Listbox.tsx',
+    componentOrigin:
+      'Dit component is overgenomen van de Gemeente Utrecht met styling van de Rijkshuisstijl Community.',
+  },
+} as Meta<typeof Listbox>;
+
+type Story = StoryObj<typeof meta>;
+export default meta;
+
+export const Default: Story = {
+  args: {
+    children: (
+      <>
+        <ListboxOption>Option #1</ListboxOption>
+        <ListboxOption>Option #2</ListboxOption>
+        <ListboxOption>Option #3</ListboxOption>
+      </>
+    ),
+  },
+};